+++ apps/opencs/view/widget/colorpickerpopup.cpp
@@ -4,14 +4,14 @@
 #include <QPushButton>
 #include <QEvent>
 #include <QKeyEvent>
-#include <QLayout>
-#include <QStyleOption>
+#include <QStyleOptionButton>
+#include <QVBoxLayout>
 
 CSVWidget::ColorPickerPopup::ColorPickerPopup(QWidget *parent) 
     : QFrame(parent)
 {
     setWindowFlags(Qt::Popup);
-    setFrameStyle(QFrame::Box | QFrame::Plain);
+    setFrameStyle(QFrame::Box | static_cast<int>(QFrame::Plain));
     hide();
 
     mColorPicker = new QColorDialog(this);
